requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
sqlalchemy==2.0.25
psycopg2-binary==2.9.9
alembic==1.13.1
pydantic==2.6.0
pydantic-settings==2.1.0
python-multipart==0.0.9
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
httpx==0.26.0
jinja2==3.1.3

# fastapi==0.109.0
# uvicorn==0.27.0
# sqlalchemy==2.0.25
# psycopg2-binary==2.9.9
# alembic==1.13.1            # For Database Migrations
# pydantic==2.6.0
# pydantic-settings==2.1.0   # For .env management
# python-multipart==0.0.9    # For File Uploads
# python-jose[cryptography]==3.3.0  # For JWT Auth
# passlib[bcrypt]==1.7.4     # For Password Hashing
# httpx==0.26.0              # For calling Ollama (Async)